+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## [2.0.1](https://github.com/e-Spirit/fsxa-ui/compare/v2.0.0...v2.0.1) (2021-01-15)
+
+
+### Bug Fixes
+
+* 🐛 fix responsive layout of ProductDetailSection ([#41](https://github.com/e-Spirit/fsxa-ui/issues/41)) ([0f2e7e1](https://github.com/e-Spirit/fsxa-ui/commit/0f2e7e10f47f2b6a2ccec24fbf93b9a6dd966709))
+
+# [2.0.0](https://github.com/e-Spirit/fsxa-ui/compare/v1.0.0...v2.0.0) (2021-01-15)
+
+
+* Fix mobile menu (#35) ([65ec518](https://github.com/e-Spirit/fsxa-ui/commit/65ec518312051d81d7c1a5ce01bfc62e775b30bb)), closes [#35](https://github.com/e-Spirit/fsxa-ui/issues/35)
+
+
+### Features
+
+* 🎸 Let HeaderSection use slots ([#37](https://github.com/e-Spirit/fsxa-ui/issues/37)) ([52306e3](https://github.com/e-Spirit/fsxa-ui/commit/52306e3d763d86744222c6e30ec78b14b49558e0))
+
+
+### BREAKING CHANGES
+
+* If an item is clicked a itemClicked event is emitted.
+
 # 1.0.0 (2020-11-30)